permisos y aes

+++ Practica2/AESEncription/my_ocl.h
@@ -2,6 +2,6 @@
 
 #define _OCL_H
 
-int aes_encriptionOCL(float *im, float *im_out, int height, int width);
-int aes_decriptionOCL(float *im, float *im_out, int height, int width);
+int aes_encriptionOCL(double *im, double *im_out, int height, int width);
+int aes_decriptionOCL(double *im, double *im_out, int height, int width);
 #endif
